test(e2e): run the GFD label unit tests in the normal unit job
The assertions package is //go:build e2e because it imports the
Kubernetes client. gfd_labels_test.go inherited that tag, so its seven
tests ran nowhere in CI: the unit job invokes go test untagged, and
make e2e-gpu-operator targets the Ginkgo suite package rather than this
subpackage's Test* funcs.

That is the same defect class the parent commit fixes. A check that
cannot fail and a check that never runs both contribute no signal.

ExpectedGFDLabels and DiffGFDLabels need nothing but the standard
library, so this splits them into an untagged gfd_labels.go and leaves
the cluster-facing WaitGFDLabels in a tagged gfd_labels_wait.go. The
test file loses its tag and now runs in the normal path.

Verified by mutation: making DiffGFDLabels return nil unconditionally,
which restores the original never-fails behaviour, turns four of the
seven tests RED under plain go test. go vet -tags e2e ./tests/...
stays clean, so the Ginkgo suite still compiles.
