+++ tests/e2e/go/assertions/gfd_labels_test.go
@@ -1,5 +1,3 @@
-//go:build e2e
-
 // Copyright 2026 NVIDIA CORPORATION
 // SPDX-License-Identifier: Apache-2.0
